+++ Tests/Specs/AnomaliesTestCase.swift
@@ -55,10 +55,51 @@ class AnomaliesTestCase: XCTestCase {
     }
     
     func testWrongAnomalyFailsTheTest() {
-        let exp = expectation(description: "Should receive item event anomaly")
+        let exp = expectation(description: "Should not receive item event anomaly")
         exp.isInverted = true
         sut.anomalyAction = exp.expect(anomaly: .itemEventCalledWithCellType(ObjectIdentifier(MemoryStorage.self)))
         sut.reportAnomaly(.weirdIndexPathAction(indexPath(0, 0)))
+        waitForExpectations(timeout: 0.1)
+    }
+    
+    func testSilencingAnomalyLeadsToAnomalyNotBeingTriggered() {
+        let exp = expectation(description: "Should receive item event anomaly")
+        exp.isInverted = true
+        sut.anomalyAction = exp.expect(anomaly: .itemEventCalledWithCellType(ObjectIdentifier(MemoryStorage.self)))
+        sut.silenceAnomaly(.itemEventCalledWithCellType(ObjectIdentifier(MemoryStorage.self)))
+        sut.reportAnomaly(.itemEventCalledWithCellType(ObjectIdentifier(MemoryStorage.self)))
+        waitForExpectations(timeout: 0.1)
+    }
+    
+    func testNotSilencedAnomaliesAreStillTriggered() {
+        let exp = expectation(description: "Should receive item event anomaly")
+        sut.anomalyAction = exp.expect(anomaly: .itemEventCalledWithCellType(ObjectIdentifier(MemoryStorage.self)))
+        sut.silenceAnomaly(.itemEventCalledWithCellType(ObjectIdentifier(BaseStorage.self)))
+        sut.reportAnomaly(.itemEventCalledWithCellType(ObjectIdentifier(MemoryStorage.self)))
+        waitForExpectations(timeout: 0.1)
+    }
+    
+    func testSilencingAnomalyUsingBlockWorks() {
+        let exp = expectation(description: "Should receive item event anomaly")
+        exp.isInverted = true
+        sut.anomalyAction = exp.expect(anomaly: .itemEventCalledWithCellType(ObjectIdentifier(MemoryStorage.self)))
+        sut.silenceAnomaly { anomaly  in
+            switch anomaly {
+            case .itemEventCalledWithCellType: return true
+            default: return false
+            }
+        }
+        sut.reportAnomaly(.itemEventCalledWithCellType(ObjectIdentifier(MemoryStorage.self)))
+        waitForExpectations(timeout: 0.1)
+    }
+    
+    func testSilencingAnomalyUsingBlock_StillTriggersOtherAnomalies() {
+        let exp = expectation(description: "Should receive item event anomaly")
+        sut.anomalyAction = exp.expect(anomaly: .itemEventCalledWithCellType(ObjectIdentifier(MemoryStorage.self)))
+        sut.silenceAnomaly { anomaly in
+            anomaly == .weirdIndexPathAction(indexPath(0, 0))
+        }
+        sut.reportAnomaly(.itemEventCalledWithCellType(ObjectIdentifier(MemoryStorage.self)))
         waitForExpectations(timeout: 0.1)
     }
 }